Sonar (#370): split config/settings.py into per-topic modules
config/settings.py was 775 lines, tripping Sonar S104's 500-line
cap. The cleanest cut without breaking `from config.settings import
X` callers is to keep settings.py as the canonical Django entry
point and host topical sub-modules alongside it that re-export their
constants. Split out:

- ``config/_channels.py`` — `_build_channel_layers()` + the
  AWS/GCP Redis TLS posture matrix (kept the existing
  underscore-prefixed name so ``tests/config/test_channel_layers``
  patches still resolve)
- ``config/_cloud.py`` — cloud provider, S3, ECS / GKE engine task
  runner, and SQS/Pub-Sub `QUEUE_CONFIG`
- ``config/_logging_config.py`` — `LOG_LEVEL` + `LOGGING` dictConfig
- ``config/_oidc_settings.py`` — authentication backends, magic-link
  config, OIDC endpoint discovery, `OIDC_EXEMPT_URLS`,
  `SESSION_COOKIE_AGE`
- ``config/_terminal_assets.py`` — `TERMINAL_CDN_ASSETS` URL/SRI map

settings.py now sits at 437 lines and re-imports each name verbatim
so the public ``config.settings`` surface is unchanged. Drive-by
fix-up: tighten `_validate_credential_spec` to return the
`SCMCredentialSpec | DeploymentProfileSpec` union so mypy stays
happy on `spec.name` access, and let ruff strip the now-unnecessary
quoted forward refs and superseded `noqa: ANN401` markers.
